fix: InputUtil 내 로직 LottoValidatorUtil로 이동
