+++ src/main/kotlin/lotto/util/LottoValidatorUtil.kt
@@ -2,6 +2,7 @@ package lotto.util
 
 import lotto.constants.Error
 import lotto.constants.LottoConstants
+import lotto.model.Lotto
 
 object LottoValidatorUtil {
 
@@ -15,6 +16,34 @@ object LottoValidatorUtil {
     fun checkNumberInRange(number: Int) {
         if (number !in LottoConstants.LOTTO_RANGE) {
             throw IllegalArgumentException(Error.EXCEPTION_MESSAGE_NOT_IN_RANGE)
+        }
+    }
+
+    fun checkMoneyAvailable(moneyString: String): Int =
+        try {
+            val money = moneyString.toInt()
+            if (money % LottoConstants.LOTTO_PRICE != 0) {
+                throw IllegalArgumentException(Error.EXCEPTION_MESSAGE_MONEY_NOT_DIVIDED_PRICE)
+            }
+            money
+        } catch (e: NumberFormatException) {
+            throw IllegalArgumentException(Error.EXCEPTION_MESSAGE_MONEY_NOT_NUMBER)
+        }
+
+    fun checkWinningNumberAvailable(winningNumberString: String): Lotto =
+        try {
+            val winningNumber = Lotto(winningNumberString.split(",").map { _numberString ->
+                val number = _numberString.toInt()
+                number
+            })
+            winningNumber
+        } catch (e: NumberFormatException) {
+            throw IllegalArgumentException(Error.EXCEPTION_MESSAGE_WINNING_NUMBER_NOT_NUMBER)
+        }
+
+    fun checkNumberOverlap(winningNumber: Lotto, number: Int) {
+        if (winningNumber.getNumbers().contains(number)) {
+            throw IllegalArgumentException(Error.EXCEPTION_MESSAGE_WINNING_NUMBER_DUPLICATED_NUMBER_EXIST)
         }
     }
 }
